adds a single point for serving ok and error response logic
