+++ internal/server/dto.go
@@ -1,6 +1,5 @@
 package server
 
-
 type ErrorRes struct {
 	Error string `json:"error"`
 }
@@ -9,4 +8,3 @@ type ErrorRes struct {
 type OkRes struct {
 	Ok bool `json:"ok"`
 }
-
