+++ internal/server/validation.go
@@ -1,7 +1,6 @@
 package server
 
 import (
-	"net/http"
 	"regexp"
 
 	"github.com/go-playground/validator"
@@ -25,15 +24,3 @@ func (cv *CustomValidator) Validate(i interface{}) error {
 
 }
 
-func (c *AppContext) validateAndBindRequest(r interface{}) error {
-
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-
-	if err := c.Validate(r); err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{err.Error()})
-	}
-
-	return nil
-}
